cos/builders.rs: colocate 2 build_cos_interface_runtime tests with production (#984 P3 Phase 4b)
Move build_cos_interface_runtime_leaves_flow_hash_seed_zero_until_promotion
and build_cos_interface_runtime_starts_exact_queue_with_zero_local_tokens
from tx/mod.rs::tests to a new mod tests in cos/builders.rs.

Adds CoSQueueConfig and FastMap imports.

cargo test --bins 865/0/2.
